Basic tests and testable apis.
The code wasn't exactly setup in a "non-cli" friendly manner. This is
fixed. loader will accept a path or a name/suite combination as an api.
This means you can call it from code.

We ensure that our tests are loaded and that we have the tests we
expect. Also, a basic run of a suite yields 10 requests to the server.
Obviously basic, but we can build from here.

Utils is responsible for setting up the test server that we want to use
and we recreate that server before every test.

Also, Mocha found a global leak in fire.js where i wasn't declared.

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>