Run script and moved the 'index' code.
Moved the index code into a runnable script. The runnable script is in
bin as per standard. This can be used within make tests to ensure that
the tests are able to run. There is no error handling as of yet so if
something fails or throws exceptions, it should return with a failing
error code rather than success.

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>